cordite-builder/requirements.txt
discord.py
Pillow